refactor: centralize HtmlTools utility
